ft/add deadline progressbar and task priority

+++ src/main/java/fr/pr70/project_pr70/back/Task.java
@@ -61,13 +61,13 @@ public class Task
         return deadline;
     }
 
-    public Double getTimePercent()
+    public double getTimePercent()
     {
         long duration = deadline.getTime() - startDate.getTime();
         Date currentDate = new Date();
         long currentPoint = currentDate.getTime() - startDate.getTime();
 
-        return (double) (currentPoint / duration);
+        return (double)currentPoint / (double)duration;
 
     }
